refactor: encapsulate the isolated SWR cache lifecycle in one hook
The cache's creation and its teardown were split across LayerProvider and the
middleware module, so the bare cleanup effect read as incidental — deleting it
left the leak-prevention silently broken. useIsolatedCacheProvider owns both,
and releaseCacheRegistration no longer needs to be public.

Also unify the two near-duplicate suspender helpers in the test into one that
can re-suspend.
