+++ src/providers/global/LayerProvider/LayerProvider.tsx
@@ -1,12 +1,12 @@
-import { type PropsWithChildren, useCallback, useEffect, useMemo, useState } from 'react'
+import { type PropsWithChildren, useMemo } from 'react'
 import { SWRConfig } from 'swr'
 
 import { type LayerThemeConfig } from '@internal-types/shared/layerContext'
 import type { LayerEvent } from '@schemas/common/layerEvents'
 import { type LayerError } from '@utils/shared/api/errorHandler'
 import { DEFAULT_LOCALE, type SupportedLocale } from '@utils/shared/i18n/supportedLocale'
-import { cacheRegistrationMiddleware, releaseCacheRegistration } from '@utils/shared/swr/cacheRegistrationMiddleware'
 import { DEFAULT_SWR_CONFIG } from '@utils/shared/swr/defaultSWRConfig'
+import { cacheRegistrationMiddleware, useIsolatedCacheProvider } from '@utils/shared/swr/isolatedSWRCache'
 import { localeKeyMiddleware } from '@utils/shared/swr/localeKeyMiddleware'
 import { AuthInputProvider } from '@providers/global/AuthInput/AuthInputProvider'
 import type { Environment, EnvironmentConfigOverride } from '@providers/global/Environment/environmentConfigs'
@@ -49,14 +49,11 @@ export const LayerProvider = ({
   usePlaidSandbox,
   ...restProps
 }: PropsWithChildren<LayerProviderProps>) => {
-  const [cache] = useState(() => new Map())
-  const provider = useCallback(() => cache, [cache])
+  const provider = useIsolatedCacheProvider()
   const swrConfig = useMemo(
     () => ({ ...DEFAULT_SWR_CONFIG, use: [cacheRegistrationMiddleware, localeKeyMiddleware], provider }),
     [provider],
   )
-
-  useEffect(() => () => releaseCacheRegistration(cache), [cache])
 
   let environment: Environment | undefined
   let environmentConfigOverride: EnvironmentConfigOverride | undefined
